trace/cloud-trace-demo-app-opentelemetry/app/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-propagator-gcp==1.3.0
opentelemetry-instrumentation-flask==0.35b0
opentelemetry-instrumentation-requests==0.35b0
google-cloud-trace==1.7.3
google-cloud-logging==3.3.0
opentelemetry-api~=1.0
opentelemetry-exporter-gcp-monitoring~=1.1.0a0


